// src/types.ts
export interface MiniUser {
  id: string;
  username: string;
  displayName: string;
  profilePic: string;
}
// ...
export interface ReelComment {
  id: string;
  username: string;
  userDisplayName: string;
  profilePic: string;
  text: string;
  timestamp: string;
}

export interface Reel {
  id: string;
  user: MiniUser;
  videoUrl: string;
  aspectRatio: '9:16';
  caption: string;
  hashtags: string[];
  soundTitle: string;
  likes: string[]; // User IDs who liked
  comments: ReelComment[];
  shares: number;
  saves: string[]; // User IDs who saved
  views: number;
  durationSeconds: number;
  repostsCount?: number;
  repliesCount?: number;
  downloadsCount?: number;
  allowDownloads?: boolean;
  boosts?: any;
}